Add fetchFavories into Login container
